move generate as well
