+++ mlx_lm/__init__.py
@@ -7,4 +7,5 @@ from ._version import __version__
 os.environ["TRANSFORMERS_NO_ADVISORY_WARNINGS"] = "1"
 
 from .convert import convert
-from .utils import generate, load, stream_generate
+from .generate import generate, stream_generate
+from .utils import load
